requirements.txt
bitly-api==0.3
click==5.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.6
django-bootstrap3==6.2.2
django-debug-toolbar==1.4
django-extra-views==0.7.1
django-rq==0.8.0
django-widget-tweaks==1.4.1
gunicorn==19.3.0
oauthlib==1.0.3
python-twitter==2.2
redis==2.10.5
requests==2.8.1
requests-oauthlib==0.5.0
rq==0.5.6
rq-scheduler==0.5.1
six==1.10.0
sqlparse==0.1.18
static3==0.6.1
Wand==0.4.1
wheel==0.24.0
psycopg2
